Try exacting clipper-related data to own type
This commit adds a new type that will make `clipperActivated`
unnecessary, as the presence of a `Just ClipperModule` in the main state
(Model) signifies that module got activated.

`updateModel` seems to have a lot of responsibilities, this might need
to change. Maybe extracting some of the logic to their own events?

Side note: The code style hasn't run through the linter yet so some
indents are still wrong.  This will be changed later on.